added in IO support, minor bugs fixed
