+++ matrixprofile/__init__.py
@@ -5,6 +5,7 @@ from matrixprofile.compute import compute
 from matrixprofile.visualize import visualize
 from matrixprofile.analyze import analyze
 from matrixprofile import discover
+from matrixprofile import transform
 from matrixprofile import utils
 from matrixprofile import io
 from matrixprofile import algorithms
