+++ tests/test_annotation_vector.py
@@ -9,15 +9,47 @@ range = getattr(__builtins__, 'xrange', range)
 # end of py2 compatability boilerplate
 
 import os
+import tempfile
 
 import pytest
 
 import numpy as np
 
-from matrixprofile import utils
+from matrixprofile import transform
 from matrixprofile import compute
+from matrixprofile import io
 
 import matrixprofile
+
+
+def test_av_io():
+    ts = np.random.uniform(size=1024)
+    w = 32
+
+    profile = compute(ts, w)
+    profile = transform.apply_av(profile, "default")
+
+    out = os.path.join(tempfile.gettempdir(), 'mp.json')
+    io.to_disk(profile, out)
+
+    dprofile = io.from_disk(out)
+
+    keys = set(profile.keys())
+    keysb = set(dprofile.keys())
+
+    assert(keys == keysb)
+
+    # check values same
+    for k, v in profile.items():
+        if isinstance(v, np.ndarray):
+            np.testing.assert_equal(v, dprofile[k])
+        elif k == 'data':
+            pass
+        else:
+            assert(v == dprofile[k])
+
+    np.testing.assert_equal(profile['data']['ts'], dprofile['data']['ts'])
+    np.testing.assert_equal(profile['data']['query'], dprofile['data']['query'])
 
 
 def test_default_valid():
@@ -25,7 +57,7 @@ def test_default_valid():
     expect = [1, 1, 1, 1]
     w = 3
 
-    av = utils.make_default_av(ts_arr, w)
+    av = transform.make_default_av(ts_arr, w)
 
     assert(len(av) == len(expect))
 
@@ -42,7 +74,7 @@ def test_complexity_valid():
     w = 3
 
     for i in range(len(ts_arr)):
-        av = utils.make_complexity_av(ts_arr[i], w)
+        av = transform.make_complexity_av(ts_arr[i], w)
 
         assert(len(av) == len(expect[i]))
 
@@ -59,7 +91,7 @@ def test_meanstd_valid():
     w = 3
 
     for i in range(len(ts_arr)):
-        av = utils.make_meanstd_av(ts_arr[i], w)
+        av = transform.make_meanstd_av(ts_arr[i], w)
 
         assert(len(av) == len(expect[i]))
 
@@ -76,7 +108,7 @@ def test_clipping_valid():
     w = 3
 
     for i in range(len(ts_arr)):
-        av = utils.make_clipping_av(ts_arr[i], w)
+        av = transform.make_clipping_av(ts_arr[i], w)
 
         assert(len(av) == len(expect[i]))
 
@@ -85,68 +117,68 @@ def test_clipping_valid():
 
 def test_default_invalid():
     with pytest.raises(ValueError) as excinfo:
-        utils.make_default_av("array", 3)
+        transform.make_default_av("array", 3)
         assert 'make_default_av expects ts to be array-like' \
             in str(excinfo.value)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.make_default_av([[1, 2, 3], [1, 2, 3]], 3)
+        transform.make_default_av([[1, 2, 3], [1, 2, 3]], 3)
         assert 'make_default_av expects ts to be one-dimensional' \
             in str(excinfo.value)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.make_default_av([1, 2, 3], "window")
+        transform.make_default_av([1, 2, 3], "window")
         assert 'make_default_av expects window to be an integer' \
             in str(excinfo.value)
 
 
 def test_complexity_invalid():
     with pytest.raises(ValueError) as excinfo:
-        utils.make_complexity_av("array", 3)
+        transform.make_complexity_av("array", 3)
         assert 'make_complexity_av expects ts to be array-like' \
             in str(excinfo.value)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.make_complexity_av([[1, 2, 3], [1, 2, 3]], 3)
+        transform.make_complexity_av([[1, 2, 3], [1, 2, 3]], 3)
         assert 'make_complexity_av expects ts to be one-dimensional' \
             in str(excinfo.value)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.make_complexity_av([1, 2, 3], "window")
+        transform.make_complexity_av([1, 2, 3], "window")
         assert 'make_complexity_av expects window to be an integer' \
             in str(excinfo.value)
 
 
 def test_meanstd_invalid():
     with pytest.raises(ValueError) as excinfo:
-        utils.make_meanstd_av("array", 3)
+        transform.make_meanstd_av("array", 3)
         assert 'make_meanstd_av expects ts to be array-like' \
             in str(excinfo.value)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.make_meanstd_av([[1, 2, 3], [1, 2, 3]], 3)
+        transform.make_meanstd_av([[1, 2, 3], [1, 2, 3]], 3)
         assert 'make_meanstd_av expects ts to be one-dimensional' \
             in str(excinfo.value)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.make_meanstd_av([1, 2, 3], "window")
+        transform.make_meanstd_av([1, 2, 3], "window")
         assert 'make_meanstd_av expects window to be an integer' \
             in str(excinfo.value)
 
 
 def test_clipping_invalid():
     with pytest.raises(ValueError) as excinfo:
-        utils.make_clipping_av("array", 3)
+        transform.make_clipping_av("array", 3)
         assert 'make_clipping_av expects ts to be array-like' \
             in str(excinfo.value)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.make_clipping_av([[1, 2, 3], [1, 2, 3]], 3)
+        transform.make_clipping_av([[1, 2, 3], [1, 2, 3]], 3)
         assert 'make_clipping_av expects ts to be one-dimensional' \
             in str(excinfo.value)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.make_clipping_av([1, 2, 3], "window")
+        transform.make_clipping_av([1, 2, 3], "window")
         assert 'make_clipping_av expects window to be an integer' \
             in str(excinfo.value)
 
@@ -158,9 +190,9 @@ def test_apply_default_av_valid():
     profile = compute(ts, windows=w)
     expect = profile['mp']
 
-    profile = utils.apply_av(profile, "default")
+    profile = transform.apply_av(profile, "default")
 
-    np.testing.assert_almost_equal(profile['mp'], expect)
+    np.testing.assert_almost_equal(profile['cmp'], expect)
 
 
 def test_apply_complexity_av_valid():
@@ -170,9 +202,9 @@ def test_apply_complexity_av_valid():
     profile = compute(ts, windows=w)
     expect = profile['mp'] * 2
 
-    profile = utils.apply_av(profile, "complexity")
+    profile = transform.apply_av(profile, "complexity")
 
-    np.testing.assert_almost_equal(profile['mp'], expect)
+    np.testing.assert_almost_equal(profile['cmp'], expect)
 
 
 def test_apply_meanstd_av_valid():
@@ -182,9 +214,9 @@ def test_apply_meanstd_av_valid():
     profile = compute(ts, windows=w)
     expect = profile['mp'] * 2
 
-    profile = utils.apply_av(profile, "meanstd")
+    profile = transform.apply_av(profile, "meanstd")
 
-    np.testing.assert_almost_equal(profile['mp'], expect)
+    np.testing.assert_almost_equal(profile['cmp'], expect)
 
 
 def test_apply_clipping_av_valid():
@@ -194,9 +226,9 @@ def test_apply_clipping_av_valid():
     profile = compute(ts, windows=w)
     expect = profile['mp'] * 2
 
-    profile = utils.apply_av(profile, "clipping")
+    profile = transform.apply_av(profile, "clipping")
 
-    np.testing.assert_almost_equal(profile['mp'], expect)
+    np.testing.assert_almost_equal(profile['cmp'], expect)
 
 
 def test_apply_custom_av_valid():
@@ -207,9 +239,9 @@ def test_apply_custom_av_valid():
     expect = profile['mp'] * 2
 
     av = [0., 0., 0., 0.]
-    profile = utils.apply_av(profile, "custom", av)
+    profile = transform.apply_av(profile, "custom", av)
 
-    np.testing.assert_almost_equal(profile['mp'], expect)
+    np.testing.assert_almost_equal(profile['cmp'], expect)
 
 
 def test_apply_av_invalid():
@@ -217,34 +249,28 @@ def test_apply_av_invalid():
     w = 3
 
     with pytest.raises(ValueError) as excinfo:
-        utils.apply_av("profile", "default")
+        transform.apply_av("profile", "default")
         assert 'apply_av expects profile as an MP data structure' \
             in str(excinfo.value)
 
     profile = compute(ts, windows=w)
 
     with pytest.raises(ValueError) as excinfo:
-        utils.apply_av(profile, "custom", "av")
+        transform.apply_av(profile, "custom", "av")
         assert 'apply_av expects custom_av to be array-like' \
             in str(excinfo.value)
 
-    profile = compute(ts, windows=w)
-
     with pytest.raises(ValueError) as excinfo:
-        utils.apply_av(profile, "not a parameter")
+        transform.apply_av(profile, "not a parameter")
         assert 'av parameter is invalid' \
             in str(excinfo.value)
 
-    profile = compute(ts, windows=w)
-
     with pytest.raises(ValueError) as excinfo:
-        utils.apply_av(profile, "custom", [0.9, 0.9, 0.9])
+        transform.apply_av(profile, "custom", [0.9, 0.9, 0.9])
         assert 'Lengths of annotation vector and mp are different' \
             in str(excinfo.value)
 
-    profile = compute(ts, windows=w)
-
     with pytest.raises(ValueError) as excinfo:
-        utils.apply_av(profile, "custom", [0.5, 0.5, 0.6, 1.2, -0.4])
+        transform.apply_av(profile, "custom", [0.5, 0.5, 0.6, 1.2, -0.4])
         assert 'Annotation vector values must be between 0 and 1' \
             in str(excinfo.value)
